Contact Card refactor
